Initial demprep conversion + gtests (#4158)
