--- isis/src/base/apps/demprep/tsts/prepinside/Makefile
@@ -1,10 +1,0 @@
-APPNAME = demprep
-
-labels.txt.IGNORELINES = Bytes StartByte ByteOrder TileSamples TileLines
-
-include $(ISISROOT)/make/isismake.tsts
-
-commands:
-	$(APPNAME) from=$(INPUT)/ulcn2005_lpo_inside.cub \
-	  to=$(OUTPUT)/ulcn2005_lpo_inside_pad.cub > /dev/null;
-	  catlab from=$(OUTPUT)/ulcn2005_lpo_inside_pad.cub to=$(OUTPUT)/labels.txt > /dev/null;
